# docs/swagger.yaml
components:
  schemas:
    CasTask:
      properties:
        task_id:
          format: uuid4
          title: Task Id
          type: string
        task_status:
          title: Task Status
          type: string
      required:
      - task_id
      - task_status
      title: CasTask
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: Customer Analysis Service API
  version: 1.0.0
openapi: 3.0.2
paths:
  /api/v1/analysis/interests/comments:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0433\u0440\u0443\
        \u043F\u043F\u0438\u0440\u043E\u0432\u043A\u0443 \u043F\u043E \u0432\u0441\
        \u0435\u043C \u0438\u043D\u0442\u0435\u0440\u0435\u0441\u0430\u043C \u043A\
        \u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0442\u043E\u0440\u043E\u0432 \u043F\
        \u0440\u043E\u0434\u0443\u043A\u0442\u0430."
      operationId: customer_interest_comments_api_v1_analysis_interests_comments_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Commentators interests
      tags:
      - analysis_interests
  /api/v1/analysis/interests/reviews:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0433\u0440\u0443\
        \u043F\u043F\u0438\u0440\u043E\u0432\u043A\u0443 \u043F\u043E \u0432\u0441\
        \u0435\u043C \u0438\u043D\u0442\u0435\u0440\u0435\u0441\u0430\u043C \u0440\
        \u0435\u0432\u044C\u044E\u0435\u0440\u043E\u0432 \u043F\u0440\u043E\u0434\u0443\
        \u043A\u0442\u0430."
      operationId: customer_interests_reviews_api_v1_analysis_interests_reviews_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reviewers interests
      tags:
      - analysis_interests
  /api/v1/analysis/sentiments/comments/category:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u0435\u043D\
        \u0442\u0438\u043C\u0435\u043D\u0442 \u0430\u043D\u0430\u043B\u0438\u0437\
        \ \u043A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0442\u043E\u0440\u043E\
        \u0432 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u0430 \u0441\u0433\u0440\
        \u0443\u043F\u043F\u0438\u0440\u043E\u0432\u0430\u043D\u043D\u044B\u0439 \u043F\
        \u043E \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u044F\u043C"
      operationId: customer_sentiment_comments_category_api_v1_analysis_sentiments_comments_category_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Commentators sentiments by category
      tags:
      - analysis_sentiments
  /api/v1/analysis/sentiments/comments/region:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u0435\u043D\
        \u0442\u0438\u043C\u0435\u043D\u0442 \u0430\u043D\u0430\u043B\u0438\u0437\
        \ \u043A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0442\u043E\u0440\u043E\
        \u0432 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u0430 \u0441\u0433\u0440\
        \u0443\u043F\u043F\u0438\u0440\u043E\u0432\u0430\u043D\u043D\u044B\u0439 \u043F\
        \u043E \u0440\u0435\u0433\u0438\u043E\u043D\u0430\u043C"
      operationId: customer_sentiment_comments_region_api_v1_analysis_sentiments_comments_region_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Commentators sentiments by region
      tags:
      - analysis_sentiments
  /api/v1/analysis/sentiments/reviews/category:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u0435\u043D\
        \u0442\u0438\u043C\u0435\u043D\u0442 \u0430\u043D\u0430\u043B\u0438\u0437\
        \ \u0440\u0435\u0432\u044C\u044E\u0435\u0440\u043E\u0432 \u043F\u0440\u043E\
        \u0434\u0443\u043A\u0442\u0430 \u0441\u0433\u0440\u0443\u043F\u043F\u0438\u0440\
        \u043E\u0432\u0430\u043D\u043D\u044B\u0439 \u043F\u043E \u043A\u0430\u0442\
        \u0435\u0433\u043E\u0440\u0438\u044F\u043C"
      operationId: customer_sentiment_reviews_category_api_v1_analysis_sentiments_reviews_category_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reviewers sentiments by category
      tags:
      - analysis_sentiments
  /api/v1/analysis/sentiments/reviews/region:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u0435\u043D\
        \u0442\u0438\u043C\u0435\u043D\u0442 \u0430\u043D\u0430\u043B\u0438\u0437\
        \ \u0440\u0435\u0432\u044C\u044E\u0435\u0440\u043E\u0432 \u043F\u0440\u043E\
        \u0434\u0443\u043A\u0442\u0430 \u0441\u0433\u0440\u0443\u043F\u043F\u0438\u0440\
        \u043E\u0432\u0430\u043D\u043D\u044B\u0439 \u043F\u043E \u0440\u0435\u0433\
        \u0438\u043E\u043D\u0430\u043C"
      operationId: customer_sentiment_reviews_region_api_v1_analysis_sentiments_reviews_region_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reviewers sentiments by region
      tags:
      - analysis_sentiments
  /api/v1/analysis/similarity/comments/reputation:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0430\u043D\u0430\
        \u043B\u0438\u0437 \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u0438 \u043E\
        \u0442\u0437\u044B\u0432\u043E\u0432 \u043A\u0430\u0436\u0434\u043E\u0433\u043E\
        \ \u043A\u043B\u0438\u0435\u043D\u0442\u0430 \u0441\u043E \u0437\u043D\u0430\
        \u0447\u0435\u043D\u0438\u0435\u043C \u0440\u0435\u043F\u0443\u0442\u0430\u0446\
        \u0438\u0438."
      operationId: get_customer_similarity_analysis_product_by_comments_api_v1_analysis_similarity_comments_reputation_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Similarity comments customer analysis by product
      tags:
      - analysis_similarity
  /api/v1/analysis/similarity/reviews/reputation:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0430\u043D\u0430\
        \u043B\u0438\u0437 \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u0438 \u043E\
        \u0442\u0437\u044B\u0432\u043E\u0432 \u043A\u0430\u0436\u0434\u043E\u0433\u043E\
        \ \u043A\u043B\u0438\u0435\u043D\u0442\u0430 \u0441\u043E \u0437\u043D\u0430\
        \u0447\u0435\u043D\u0438\u0435\u043C \u0440\u0435\u043F\u0443\u0442\u0430\u0446\
        \u0438\u0438."
      operationId: get_customer_similarity_analysis_product_by_reviews_api_v1_analysis_similarity_reviews_reputation_get
      parameters:
      - in: query
        name: product_name_id
        required: true
        schema:
          title: Product Name Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Similarity reviews customer analysis by product
      tags:
      - analysis_similarity
...